Implement S3 storage on top of awk-sdk-s3
Previously we were just doing PUT requests directly against the S3
REST API and relying on the bucket policy to give access to the right
machines.  However, when you send a PUT without an Authorization
header, the objects get created with owner "anonymous".  When you try
to access these objects from the S3 console you get access denied,
because you're logged in and therefore not anonymous, or something.