Record treatment batch IDs and summarise treatments on the assessment view
Currently the batch IDs aren't really doing anything except providing
a historic record of which treatments were created from the same
batch.  It turns out the relationship tables had enough information to
link between treatments and assessments, so that seems to be enough to
create the treatment summary view needed.

Still need to get the treatment summary showing up in the CSV view,
but will confirm that this is on the right track first.